create keyboard_thread.py
This is first attempt at using threads to organize multiple keypresses.

Need to test on hardware. Verify .05 sleep allows striker to reach end position. There is a moving state that can be read from the controller.

There is an issue where the main thread gets stuck and will not see the "exit" key. Printing to the terminal seems to increase the occurrence of issue.